Move waiting list related tests to own directory
